requirements.txt
# Core dependencies
click>=8.1.0
rich>=13.0.0
pdfplumber>=0.10.0
sqlalchemy>=2.0.0
pandas>=2.1.0
python-dateutil>=2.8.0
unidecode>=1.3.0
pyyaml>=6.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
